docker/security-data-loader/requirements.txt
certifi==2020.4.5.1
elasticsearch==7.6.0
pytz==2020.1
urllib3==1.25.9
python-dateutil==2.8.2
